Move SolutionIndicators struct to own compilation unit.
